--- api/_config/index.ts
@@ -1,2 +1,0 @@
-export * from "./notionClient";
-export * from "./notionKeys";
