--- api/_config/notion/index.ts
@@ -1,3 +1,0 @@
-export * from "./types";
-export * from "./constants";
-export * from "./utils";
