--- api/_config/notionKeys.ts
@@ -1,4 +1,0 @@
-export const notionDBKeys = {
-  events: process.env.NOTION_EVENTS_DB_ID || "",
-  members: process.env.NOTION_MEMBERS_DB_ID || "",
-} as const;
