+++ api/constants.ts
@@ -1,6 +1,9 @@
 import type { UsedNotionPropertyType } from "./types";
 
-// 型から値を抽出してenum的に使用できる定数オブジェクト
+/**
+ * 型から値を抽出してenum的に使用できる定数オブジェクト
+ * データモデルのスキーマを定義する際に利用する。
+ */
 export const NOTION_PROPERTY_TYPES = {
   TITLE: "title",
   RICH_TEXT: "rich_text",
@@ -13,5 +16,9 @@ export const NOTION_PROPERTY_TYPES = {
   CHECKBOX: "checkbox",
   URL: "url",
   EMAIL: "email",
-  PHONE_NUMBER: "phone_number",
 } as const satisfies Record<string, UsedNotionPropertyType>;
+
+export const NOTION_DB_KEYS = {
+  events: process.env.NOTION_EVENTS_DB_ID || "",
+  members: process.env.NOTION_MEMBERS_DB_ID || "",
+} as const;
