+++ api/utils/createPropertyDefinition.ts
@@ -2,9 +2,15 @@ import type {
   PropertyDefinition,
   PropertyValueType,
   UsedNotionPropertyType,
-} from "./types";
+} from "../types";
 
-// 型安全なスキーマ作成用のヘルパー関数
+/**
+ * 型安全なスキーマ作成用のヘルパー関数
+ * @param name プロパティ名
+ * @param type プロパティ型
+ * @param options オプション
+ * @returns プロパティ定義
+ */
 export function createPropertyDefinition<T extends UsedNotionPropertyType>(
   name: string,
   type: T,
