+++ app/events/_components/EventListItem.tsx
@@ -1,8 +1,8 @@
+import type { Event } from "@/app/events/_api/model";
 import clsx from "clsx";
 import { DateTime } from "luxon";
 import Image from "next/image";
 import Link from "next/link";
-import type { Event } from "../ _api/model";
 
 interface EventListItemProps {
   event: Event;
